+++ clickhouse-data/src/main/java/com/clickhouse/data/value/array/ClickHouseFloatArrayValue.java
@@ -497,7 +497,7 @@ public class ClickHouseFloatArrayValue extends ClickHouseObjectValue<float[]> im
     public ClickHouseArraySequence allocate(int length, Class<?> clazz, int level) {
         if (length < 1) {
             resetToDefault();
-        } else if (length() != length) {
+        } else {
             set(new float[length]);
         }
         return this;