+++ clickhouse-data/src/main/java/com/clickhouse/data/value/array/ClickHouseShortArrayValue.java
@@ -583,7 +583,7 @@ public class ClickHouseShortArrayValue extends ClickHouseObjectValue<short[]> im
     public ClickHouseArraySequence allocate(int length, Class<?> clazz, int level) {
         if (length < 1) {
             resetToDefault();
-        } else if (length() != length) {
+        } else {
             set(new short[length]);
         }
         return this;